Fix a bug in the GLSL120 shaders.
The shaders go sequentially through the light source array. But the enabled
lights aren't necessarily in front of the array, hence we can't just loop
over lightsNumber (number of enabled lights, e.g. 3). We have to loop through
all possible lights (max 8 in classic OpenGL), and check for each if it's in an
'enabled' uniform array filled in by the application. Otherwise, disabling a
light could also disable others in the way, or disable the wrong one.