Fix a bug in the GLSL330 shaders.
The light direction computation for directional lights was wrong. It was computed
as (vertex pos - light pos), instead of the... light direction. This bug was making
the directional light act as a plain, no-attenuated point light.

It wasn't detected in a long time though, as the combination of the tests' light's
position and orientation (unfortunately) made a correct lighting for the directional
light.

But moving the directional light revealed the bug, as the light effect orientation
changed, whereas the position of a directional light shouldn't affect it, given these
lights are supposed to be infinitely far away in a fixed direction.

To follow the old OpenGL system, the position of a directional light passed to the
shaders is actually the light direction. Though it may change in the future since
this can be rather confusing.