Fix some lighting issues.
In GLSL330: directional light and spot light directions were wrong.
A combination of unfortunate positions *and* orientations was making them appear
accurate during the tests, but changing their position and/or position proved them
wrong. The spot direction calculation was right for the default orientation, but
wrong otherwise.

Another bug was with GLSL120 shaders not passing through all possible lights. The
problem is that OpenGL lights don't have fixed IDs: they do not necessarily stay
in the same order when one disables/enables them, but are more likely to "roll"
in the gl_LightSource array. Since the shaders were sequentially processing only
the first "lightsNumber" (number of lights we want the shader to use) lights, this
resulted in lights surprisingly disabled when disabling another, and multiple
lights disabled at once. To fix this, introduce a "lightsEnabled" uniform
containing the IDs of all enabled lights. Shaders now loop through the entire
gl_LightSource array and compare the ID of each light with all the IDs in the
lightsEnabled array. It only adds the light color if its ID is in the array.

The shader part of these bugs has been fixed in previous commits and all should be
good now (let's hope so).

Create separated private functions setupFixedPipelineLights and
setupProgrammableRendererLights as the fixed pipeline needs a little more work now