Version 2.0
No new features but a change of approach that changes a lot from the previous ones.
Copying the data/ folder of the addon is so cumbersome, not talking of the double
space taken by the shaders if one forgets to delete it in the original addon folder.

Decision has thus been made to instead incorporate the minified version of shader
sources directly into the addon code and load it internly. No more need for additional
extern shader files.

This approach simplifies a lot the installation and use process.

- Update the README

- Add the ofxaddons thumbnail image