Add the first version of displacement shader. Only in GLSL120
... for now. Also come back to an approach closer to the Lighting
one. For the Mapping module, I first thought moving to a
"1 shader = 1 class"-system was a good idea, since mapping shaders
can be pretty different in their implementation and use. But that
shouldn't be necessary. By using a wider vocabulary in the
MappingShader class, it should be possible to do just any mapping
with the same class, while keeping an understandable code.