Forbid calls to Shader::reload() in child shaders (make it private).
The reload() function retrieves and then load a shader program on
thr GPU, and that was done with basically every operation changing
the shaders.

Instead, now use an internal boolean asking for reload that the base
Shader reads on begin(), and reloads if needed. This way
we reload the shader only once by frame.